chore: create view wall fixture
